add sprint burndown functionality
what: the sprint burndown class has been added which records hours and sums up the numbers for the ideal and the actual plot. functionality to set the number of weeks in a sprint and provide a number of isses override has also been added

why: the sprint burndown will help provide teams with a sense of completion / accomplishment and would be shared by the end of the sprint in the review. in addition to that, the functionality to set the number of weeks for the sprint was added to determine the start date of the sprint, as Github does not have a start date definition for milestone objects. the requirement to set the number of issues is also required because the total count of the issues to be analyzed is not known beforehand, so the user can ovverride that count

impact: high